requirements.txt
fastapi==0.99.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
uvicorn==0.22.0
requests==2.31.0
